Create PayoutDetailsBankDetails subcomponent
