+++ src/forms/PayoutDetailsForm/PayoutDetailsForm.css
@@ -105,3 +105,7 @@
     cursor: pointer;
   }
 }
+
+.bankDetailsStripeField p {
+  @apply --marketplaceH4FontStyles;
+}
