+++ src/forms/PayoutDetailsForm/PayoutDetailsFormCompany.js
@@ -2,18 +2,13 @@ import React from 'react';
 import { bool, object, string } from 'prop-types';
 import { compose } from 'redux';
 import { FormattedMessage, injectIntl, intlShape } from 'react-intl';
-import { StripeBankAccountTokenInputField, FieldTextInput } from '../../components';
+import { FieldTextInput } from '../../components';
 import * as validators from '../../util/validators';
 
 import PayoutDetailsAddress from './PayoutDetailsAddress';
+import PayoutDetailsBankDetails from './PayoutDetailsBankDetails';
 import PayoutDetailsPersonalDetails from './PayoutDetailsPersonalDetails';
-import { stripeCountryConfigs } from './PayoutDetailsForm';
 import css from './PayoutDetailsForm.css';
-
-const countryCurrency = countryCode => {
-  const country = stripeCountryConfigs(countryCode);
-  return country.currency;
-};
 
 const PayoutDetailsFormCompanyComponent = ({ fieldRenderProps }) => {
   const { disabled, form, intl, values } = fieldRenderProps;
@@ -38,11 +33,6 @@ const PayoutDetailsFormCompanyComponent = ({ fieldRenderProps }) => {
       id: 'PayoutDetailsForm.companyTaxIdRequired',
     })
   );
-
-  // StripeBankAccountTokenInputField handles the error messages
-  // internally, we just have to make sure we require a valid token
-  // out of the field. Therefore the empty validation message.
-  const bankAccountRequired = validators.required(' ');
 
   return (
     <React.Fragment>
@@ -86,19 +76,7 @@ const PayoutDetailsFormCompanyComponent = ({ fieldRenderProps }) => {
             companyAddress
           />
 
-          <div className={css.sectionContainer}>
-            <h3 className={css.subTitle}>
-              <FormattedMessage id="PayoutDetailsForm.bankDetails" />
-            </h3>
-            <StripeBankAccountTokenInputField
-              disabled={disabled}
-              name="bankAccountToken"
-              formName="PayoutDetailsForm"
-              country={country}
-              currency={countryCurrency(country)}
-              validate={bankAccountRequired}
-            />
-          </div>
+          <PayoutDetailsBankDetails country={country} disabled={disabled} />
 
           <PayoutDetailsPersonalDetails
             intl={intl}
