+++ src/forms/PayoutDetailsForm/PayoutDetailsFormIndividual.js
@@ -1,28 +1,15 @@
 import React from 'react';
 import { bool, object } from 'prop-types';
 import { compose } from 'redux';
-import { FormattedMessage, injectIntl, intlShape } from 'react-intl';
-import { StripeBankAccountTokenInputField } from '../../components';
-import * as validators from '../../util/validators';
+import { injectIntl, intlShape } from 'react-intl';
 
 import PayoutDetailsAddress from './PayoutDetailsAddress';
+import PayoutDetailsBankDetails from './PayoutDetailsBankDetails';
 import PayoutDetailsPersonalDetails from './PayoutDetailsPersonalDetails';
-import { stripeCountryConfigs } from './PayoutDetailsForm';
-import css from './PayoutDetailsForm.css';
-
-const countryCurrency = countryCode => {
-  const country = stripeCountryConfigs(countryCode);
-  return country.currency;
-};
 
 const PayoutDetailsFormIndividualComponent = ({ fieldRenderProps }) => {
   const { disabled, form, intl, values } = fieldRenderProps;
   const { country } = values;
-
-  // StripeBankAccountTokenInputField handles the error messages
-  // internally, we just have to make sure we require a valid token
-  // out of the field. Therefore the empty validation message.
-  const bankAccountRequired = validators.required(' ');
 
   return (
     <React.Fragment>
@@ -33,20 +20,7 @@ const PayoutDetailsFormIndividualComponent = ({ fieldRenderProps }) => {
         country={country}
       />
       <PayoutDetailsAddress country={country} intl={intl} disabled={disabled} form={form} />
-
-      <div className={css.sectionContainer}>
-        <h3 className={css.subTitle}>
-          <FormattedMessage id="PayoutDetailsForm.bankDetails" />
-        </h3>
-        <StripeBankAccountTokenInputField
-          disabled={disabled}
-          name="bankAccountToken"
-          formName="PayoutDetailsForm"
-          country={country}
-          currency={countryCurrency(country)}
-          validate={bankAccountRequired}
-        />
-      </div>
+      <PayoutDetailsBankDetails country={country} disabled={disabled} />
     </React.Fragment>
   );
 };
